refactor(markdown): keep files under size baselines; bump Settings bundle
The blank-line work and the earlier IME fix pushed three files past their frozen
file-size baselines. Extract/trim to restore them, with no behavior change:

- markdownPipeline/mdastToProseMirror.ts: extract the blank-line capture helpers
  and the top-level capture loop into blankLineCapture.ts.
- markdownPipeline/serializer.ts: move blankLinesJoin into serializerHandlers.ts.
- Terminal/setupImeComposition.ts: trim the (now over-300) comment blocks.
- settings/EditorSettings.tsx: extract the Whitespace settings group into a
  self-contained WhitespaceSettings.tsx (reads the store via selectors).

The Settings-page bundle limit bumps 99 -> 101 kB: the new "Preserve blank
lines" toggle plus the WhitespaceSettings module boundary pushed the chunk ~98 B
over its old ceiling (it was already at 98.8/99 kB). pnpm check:all is green.
